fix: formatDateTime to dates.ts + use locale prop
